+++ src/gui/Onyx32.Gui/src/Controls/DateTime/DateTime.h
@@ -11,7 +11,7 @@ namespace Onyx32::Gui::Controls
 				const unsigned int height,
 				const unsigned int xPos,
 				const unsigned int yPos,
-				const unsigned int controlId);
+				const uint64_t controlId);
 			virtual ~DateTime();
 
 			virtual void Initialize() override;
@@ -19,7 +19,7 @@ namespace Onyx32::Gui::Controls
 			virtual void SetOnChange(OnDateTimeChange& onChange) override;
 			LRESULT DateTime::Process(unsigned int message, WPARAM wParam, LPARAM lParam) override;
 
-			static IDateTime* Create(IWindow* parent, unsigned int controlId, unsigned int width, unsigned int height, unsigned int xPos, unsigned int yPos);
+			static IDateTime* Create(IWindow* parent, uint64_t controlId, unsigned int width, unsigned int height, unsigned int xPos, unsigned int yPos);
 		
 		protected:
 			OnDateTimeChange& _onChange;
